# openapi.yaml
openapi: "3.1.0"
info:
  title: VocalTwist Voice Middleware API
  version: "1.0.0"
  description: |
    VocalTwist is a plug-and-play voice middleware layer that provides
    speech-to-text (STT), text-to-speech (TTS), and health/discovery endpoints.

    It is designed to be mounted as a FastAPI router inside any existing
    application, or run standalone via `VocalTwistTest/app.py`.

    ## Authentication
    API key authentication is **optional** (disabled by default for demo mode).
    When enabled, every request must include `X-API-Key: <key>`.

    ## Rate limits
    Default: 20 requests/minute on transcribe endpoints, 30/minute on speak.
    Configurable via environment variables.
servers:
  - url: http://localhost:8000
    description: Local development / demo
security: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    TranscribeResponse:
      type: object
      required: [text, display_text, language, duration_ms]
      properties:
        text:
          type: string
          description: Normalised transcript (lowercased, trimmed).
          example: "hello world"
        display_text:
          type: string
          description: Display-ready transcript (original casing from the model).
          example: "Hello world"
        language:
          type: string
          description: Detected BCP-47 language code.
          example: "en"
        duration_ms:
          type: integer
          description: Audio duration in milliseconds.
          example: 2340
    SpeakRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 4096
          description: Text to synthesise.
          example: "Hello, how can I help you?"
        voice:
          type: string
          nullable: true
          description: |
            edge-tts neural voice name, e.g. `en-US-AriaNeural`.
            Defaults to the server's `VOCALTWIST_DEFAULT_VOICE`.
          example: "en-US-AriaNeural"
        language:
          type: string
          nullable: true
          description: |
            BCP-47 language code used to auto-select a voice when `voice`
            is not specified.
          example: "en"
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "Audio file too large"
        request_id:
          type: string
          description: Correlation ID for log tracing.
          example: "req_7f3c1a2b"
paths:
  /api/transcribe:
    post:
      operationId: transcribe
      summary: Transcribe an audio file to text
      description: |
        Upload an audio file and receive a JSON transcript.
        Supported formats: WebM, WAV, MP4, OGG, MPEG, FLAC.
        Maximum file size: 10 MB (configurable).
      tags: [STT]
      security:
        - {}
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                  description: Audio file to transcribe.
                language:
                  type: string
                  description: Hint the expected language (BCP-47). Optional.
                  example: "en"
      responses:
        "200":
          description: Successful transcription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscribeResponse'
        "400":
          description: Invalid audio file or unsupported format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: API key required but missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "413":
          description: Audio file exceeds the configured size limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error (missing required field)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal transcription error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/transcribe-ambient:
    post:
      operationId: transcribeAmbient
      summary: Transcribe with forced VAD filter (ambient mode)
      description: |
        Identical to `/api/transcribe` but always enables the Silero VAD
        pre-filter. Use this endpoint from `AmbientVAD` in the frontend.
      tags: [STT]
      security:
        - {}
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                language:
                  type: string
                  example: "en"
      responses:
        "200":
          description: Successful transcription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscribeResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "429":
          $ref: '#/components/responses/TooManyRequests'
        "500":
          $ref: '#/components/responses/InternalError'
  /api/speak:
    post:
      operationId: speak
      summary: Synthesise text to MP3 audio
      description: |
        Convert text to speech using the configured TTS provider (edge-tts by
        default). Returns raw MP3 bytes (`audio/mpeg`).
      tags: [TTS]
      security:
        - {}
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeakRequest'
            example:
              text: "Hello! How can I help you today?"
              voice: "en-US-AriaNeural"
              language: "en"
      responses:
        "200":
          description: MP3 audio bytes
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
        "429":
          $ref: '#/components/responses/TooManyRequests'
        "500":
          $ref: '#/components/responses/InternalError'
...